+++ Source/CesiumEditor/Private/IonQuickAddPanel.h
@@ -2,7 +2,6 @@
 
 #pragma once
 
-#include "Dialogs/CustomDialog.h"
 #include "Widgets/SCompoundWidget.h"
 #include "Widgets/Views/STableRow.h"
 #include <string>