+++ Source/CesiumRuntime/Private/CesiumGeoreference.cpp
@@ -421,8 +421,7 @@ bool ACesiumGeoreference::SwitchToLevel(int32 Index) {
 }
 
 FVector
-ACesiumGeoreference::InaccurateGetGeoreferenceOriginLongitudeLatitudeHeight()
-    const {
+ACesiumGeoreference::GetGeoreferenceOriginLongitudeLatitudeHeight() const {
   return FVector(OriginLongitude, OriginLatitude, OriginHeight);
 }
 
@@ -455,17 +454,15 @@ void ACesiumGeoreference::SetGeoreferenceOriginEcef(
       _geoTransforms.TransformEcefToLongitudeLatitudeHeight(TargetEcef));
 }
 
-void ACesiumGeoreference::
-    InaccurateSetGeoreferenceOriginLongitudeLatitudeHeight(
-        const FVector& targetLongitudeLatitudeHeight) {
+void ACesiumGeoreference::SetGeoreferenceOriginLongitudeLatitudeHeight(
+    const FVector& targetLongitudeLatitudeHeight) {
   this->SetGeoreferenceOriginLongitudeLatitudeHeight(glm::dvec3(
       targetLongitudeLatitudeHeight.X,
       targetLongitudeLatitudeHeight.Y,
       targetLongitudeLatitudeHeight.Z));
 }
 
-void ACesiumGeoreference::InaccurateSetGeoreferenceOriginEcef(
-    const FVector& TargetEcef) {
+void ACesiumGeoreference::SetGeoreferenceOriginEcef(const FVector& TargetEcef) {
   this->SetGeoreferenceOriginEcef(
       glm::dvec3(TargetEcef.X, TargetEcef.Y, TargetEcef.Z));
 }
@@ -485,14 +482,14 @@ void ACesiumGeoreference::BeginPlay() {
     return;
   }
 
-  if (!this->WorldOriginCamera) {
+  if (!this->SubLevelCamera) {
     // Find the first player's camera manager
     APlayerController* pPlayerController = pWorld->GetFirstPlayerController();
     if (pPlayerController) {
-      this->WorldOriginCamera = pPlayerController->PlayerCameraManager;
+      this->SubLevelCamera = pPlayerController->PlayerCameraManager;
     }
 
-    if (!this->WorldOriginCamera) {
+    if (!this->SubLevelCamera) {
       UE_LOG(
           LogCesium,
           Warning,
@@ -689,13 +686,13 @@ bool ACesiumGeoreference::_updateSublevelState() {
     return false;
   }
 
-  if (!IsValid(WorldOriginCamera)) {
+  if (!IsValid(this->SubLevelCamera)) {
     return false;
   }
 
   UWorld* pWorld = this->GetWorld();
   const FIntVector& originLocation = pWorld->OriginLocation;
-  const FMinimalViewInfo& pov = this->WorldOriginCamera->ViewTarget.POV;
+  const FMinimalViewInfo& pov = this->SubLevelCamera->ViewTarget.POV;
   const FVector& cameraLocation = pov.Location;
 
   glm::dvec4 cameraAbsolute = VecMath::add4D(cameraLocation, originLocation);
@@ -729,46 +726,6 @@ bool ACesiumGeoreference::_updateSublevelState() {
 
   // activeLevel may be -1, in which case all levels will be deactivated.
   return this->SwitchToLevel(activeLevel);
-}
-
-void ACesiumGeoreference::_performOriginRebasing() {
-  UWorld* world = this->GetWorld();
-  if (!world->IsGameWorld()) {
-    return;
-  }
-  if (!IsValid(WorldOriginCamera)) {
-    return;
-  }
-  if (!this->KeepWorldOriginNearCamera) {
-    return;
-  }
-  if (this->_insideSublevel && !this->OriginRebaseInsideSublevels) {
-    // If we are not going to continue origin rebasing inside the
-    // sublevel, just set the origin back to zero if necessary,
-    // since the sublevel will be centered around zero anyways.
-    if (!world->OriginLocation.IsZero()) {
-      world->SetNewWorldOrigin(FIntVector::ZeroValue);
-    }
-    return;
-  }
-
-  // We're either not in a sublevel, or OriginRebaseInsideSublevels is true.
-  // Check whether a rebasing is necessary.
-  const FIntVector& originLocation = world->OriginLocation;
-  const FMinimalViewInfo& pov = this->WorldOriginCamera->ViewTarget.POV;
-  const FVector& cameraLocation = pov.Location;
-  bool distanceTooLarge = !cameraLocation.Equals(
-      FVector::ZeroVector,
-      this->MaximumWorldOriginDistanceFromCamera);
-  if (distanceTooLarge) {
-    // Camera has moved too far from the origin, move the origin,
-    // but make sure that no component exceeds the maximum value
-    // that can be represented as a 32bit signed integer.
-    int32 newX = clampedAdd(cameraLocation.X, originLocation.X);
-    int32 newY = clampedAdd(cameraLocation.Y, originLocation.Y);
-    int32 newZ = clampedAdd(cameraLocation.Z, originLocation.Z);
-    world->SetNewWorldOrigin(FIntVector(newX, newY, newZ));
-  }
 }
 
 void ACesiumGeoreference::_updateGeoTransforms() {
@@ -807,8 +764,6 @@ void ACesiumGeoreference::Tick(float DeltaTime) {
   if (!this->_shouldManageSubLevels()) {
     this->_insideSublevel = _updateSublevelState();
   }
-
-  _performOriginRebasing();
 }
 
 void ACesiumGeoreference::Serialize(FArchive& Ar) {
@@ -830,7 +785,7 @@ glm::dvec3 ACesiumGeoreference::TransformLongitudeLatitudeHeightToEcef(
       longitudeLatitudeHeight);
 }
 
-FVector ACesiumGeoreference::InaccurateTransformLongitudeLatitudeHeightToEcef(
+FVector ACesiumGeoreference::TransformLongitudeLatitudeHeightToEcef(
     const FVector& longitudeLatitudeHeight) const {
   glm::dvec3 ecef = this->_geoTransforms.TransformLongitudeLatitudeHeightToEcef(
       VecMath::createVector3D(longitudeLatitudeHeight));
@@ -842,7 +797,7 @@ glm::dvec3 ACesiumGeoreference::TransformEcefToLongitudeLatitudeHeight(
   return _geoTransforms.TransformEcefToLongitudeLatitudeHeight(ecef);
 }
 
-FVector ACesiumGeoreference::InaccurateTransformEcefToLongitudeLatitudeHeight(
+FVector ACesiumGeoreference::TransformEcefToLongitudeLatitudeHeight(
     const FVector& ecef) const {
   glm::dvec3 llh = this->_geoTransforms.TransformEcefToLongitudeLatitudeHeight(
       glm::dvec3(ecef.X, ecef.Y, ecef.Z));
@@ -856,7 +811,7 @@ glm::dvec3 ACesiumGeoreference::TransformLongitudeLatitudeHeightToUnreal(
       longitudeLatitudeHeight);
 }
 
-FVector ACesiumGeoreference::InaccurateTransformLongitudeLatitudeHeightToUnreal(
+FVector ACesiumGeoreference::TransformLongitudeLatitudeHeightToUnreal(
     const FVector& longitudeLatitudeHeight) const {
   glm::dvec3 ue = this->_geoTransforms.TransformLongitudeLatitudeHeightToUnreal(
       glm::dvec3(CesiumActors::getWorldOrigin4D(this)),
@@ -871,7 +826,7 @@ glm::dvec3 ACesiumGeoreference::TransformUnrealToLongitudeLatitudeHeight(
       ue);
 }
 
-FVector ACesiumGeoreference::InaccurateTransformUnrealToLongitudeLatitudeHeight(
+FVector ACesiumGeoreference::TransformUnrealToLongitudeLatitudeHeight(
     const FVector& ue) const {
   glm::dvec3 llh =
       this->_geoTransforms.TransformUnrealToLongitudeLatitudeHeight(
@@ -887,8 +842,7 @@ ACesiumGeoreference::TransformEcefToUnreal(const glm::dvec3& ecef) const {
       ecef);
 }
 
-FVector ACesiumGeoreference::InaccurateTransformEcefToUnreal(
-    const FVector& ecef) const {
+FVector ACesiumGeoreference::TransformEcefToUnreal(const FVector& ecef) const {
   glm::dvec3 ue = this->_geoTransforms.TransformEcefToUnreal(
       glm::dvec3(CesiumActors::getWorldOrigin4D(this)),
       VecMath::createVector3D(ecef));
@@ -902,8 +856,7 @@ ACesiumGeoreference::TransformUnrealToEcef(const glm::dvec3& ue) const {
       ue);
 }
 
-FVector
-ACesiumGeoreference::InaccurateTransformUnrealToEcef(const FVector& ue) const {
+FVector ACesiumGeoreference::TransformUnrealToEcef(const FVector& ue) const {
   glm::dvec3 ecef = this->_geoTransforms.TransformUnrealToEcef(
       glm::dvec3(CesiumActors::getWorldOrigin4D(this)),
       glm::dvec3(ue.X, ue.Y, ue.Z));
@@ -919,7 +872,7 @@ glm::dquat ACesiumGeoreference::TransformRotatorUnrealToEastSouthUp(
       UeLocation);
 }
 
-FRotator ACesiumGeoreference::InaccurateTransformRotatorUnrealToEastSouthUp(
+FRotator ACesiumGeoreference::TransformRotatorUnrealToEastSouthUp(
     const FRotator& UERotator,
     const FVector& ueLocation) const {
   glm::dquat q = TransformRotatorUnrealToEastSouthUp(
@@ -937,7 +890,7 @@ glm::dquat ACesiumGeoreference::TransformRotatorEastSouthUpToUnreal(
       UeLocation);
 }
 
-FRotator ACesiumGeoreference::InaccurateTransformRotatorEastSouthUpToUnreal(
+FRotator ACesiumGeoreference::TransformRotatorEastSouthUpToUnreal(
     const FRotator& EsuRotator,
     const FVector& ueLocation) const {
   glm::dquat q = TransformRotatorEastSouthUpToUnreal(
@@ -953,8 +906,8 @@ ACesiumGeoreference::ComputeEastSouthUpToUnreal(const glm::dvec3& ue) const {
       ue);
 }
 
-FMatrix ACesiumGeoreference::InaccurateComputeEastSouthUpToUnreal(
-    const FVector& ue) const {
+FMatrix
+ACesiumGeoreference::ComputeEastSouthUpToUnreal(const FVector& ue) const {
   glm::dmat3 esuToUe = this->_geoTransforms.ComputeEastSouthUpToUnreal(
       glm::dvec3(CesiumActors::getWorldOrigin4D(this)),
       glm::dvec3(ue.X, ue.Y, ue.Z));
@@ -966,8 +919,8 @@ ACesiumGeoreference::ComputeEastNorthUpToEcef(const glm::dvec3& ecef) const {
   return _geoTransforms.ComputeEastNorthUpToEcef(ecef);
 }
 
-FMatrix ACesiumGeoreference::InaccurateComputeEastNorthUpToEcef(
-    const FVector& ecef) const {
+FMatrix
+ACesiumGeoreference::ComputeEastNorthUpToEcef(const FVector& ecef) const {
   glm::dmat3 enuToEcef = this->_geoTransforms.ComputeEastNorthUpToEcef(
       glm::dvec3(ecef.X, ecef.Y, ecef.Z));
   return VecMath::createMatrix(enuToEcef);