+++ Source/CesiumRuntime/Private/CesiumMaterialUserData.cpp
@@ -14,27 +14,21 @@ void UCesiumMaterialUserData::PostEditChangeOwner() {
   if (pMaterial) {
     const FStaticParameterSet& parameters = pMaterial->GetStaticParameters();
 
-#if ENGINE_MAJOR_VERSION >= 5
-    const FMaterialLayersFunctions& layerParameters = parameters.MaterialLayers;
+#if ENGINE_MAJOR_VERSION == 5 && ENGINE_MINOR_VERSION >= 1
+    const TArray<FText>& layerNames = parameters.EditorOnly.MaterialLayers.LayerNames;
+
+    this->LayerNames.Reserve(layerNames.Num());
+
+    for (int32 i = 0; i < layerNames.Num(); ++i) {
+      this->LayerNames.Add(layerNames[i].ToString());
+    }
+#else
+    const auto& layerParameters = parameters.MaterialLayers;
 
     this->LayerNames.Reserve(layerParameters.Layers.Num());
 
     for (int32 i = 0; i < layerParameters.Layers.Num(); ++i) {
       this->LayerNames.Add(layerParameters.GetLayerName(i).ToString());
-    }
-#else
-    const TArray<FStaticMaterialLayersParameter>& layerParameters =
-        parameters.MaterialLayersParameters;
-
-    for (const FStaticMaterialLayersParameter& layerParameter :
-         layerParameters) {
-      if (layerParameter.ParameterInfo.Name != "Cesium")
-        continue;
-
-      this->LayerNames.Reserve(layerParameter.Value.LayerNames.Num());
-      for (const FText& text : layerParameter.Value.LayerNames) {
-        this->LayerNames.Add(text.ToString());
-      }
     }
 #endif
   }