+++ Source/CesiumRuntime/Public/CesiumGeoreference.h
@@ -55,17 +55,6 @@ public:
   ACesiumGeoreference();
 
   /*
-   * Whether to continue origin rebasing once inside a sublevel. If actors
-   * inside the sublevels react poorly to origin rebasing, it might be worth
-   * turning this option off.
-   */
-  UPROPERTY(
-      EditAnywhere,
-      Category = "Cesium|Cesium Sublevels",
-      meta = (EditCondition = "KeepWorldOriginNearCamera"))
-  bool OriginRebaseInsideSublevels = true;
-
-  /*
    * Whether to visualize the level loading radii in the editor. Helpful for
    * initially positioning the level and choosing a load radius.
    */
@@ -167,17 +156,6 @@ public:
   // UPROPERTY(EditAnywhere, Category = "Cesium", AdvancedDisplay)
   bool EditOriginInViewport = false;
 
-  /**
-   * If true, the world origin is periodically rebased to keep it near the
-   * camera.
-   *
-   * This is important for maintaining vertex precision in large worlds. Setting
-   * it to false can lead to jiterring artifacts when the camera gets far away
-   * from the origin.
-   */
-  UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Cesium")
-  bool KeepWorldOriginNearCamera = true;
-
 #if WITH_EDITOR
   /**
    * Places the georeference origin at the camera's current location. Rotates
@@ -194,24 +172,11 @@ public:
 #endif
 
   /**
-   * The maximum distance in centimeters that the camera may move from the
-   * world's OriginLocation before the world origin is moved closer to the
-   * camera.
+   * The camera to use to determine which sub-level is closest, so that one can
+   * be activated and all others deactivated.
    */
-  UPROPERTY(
-      EditAnywhere,
-      Category = "Cesium",
-      meta = (EditCondition = "KeepWorldOriginNearCamera", ClampMin = 0.0))
-  double MaximumWorldOriginDistanceFromCamera = 10000.0;
-
-  /**
-   * The camera to use for setting the world origin.
-   */
-  UPROPERTY(
-      EditAnywhere,
-      Category = "Cesium",
-      meta = (EditCondition = "KeepWorldOriginNearCamera"))
-  APlayerCameraManager* WorldOriginCamera;
+  UPROPERTY(EditAnywhere, Category = "CesiumSublevels")
+  APlayerCameraManager* SubLevelCamera;
 
   // TODO: Allow user to select/configure the ellipsoid.
   // Yeah, we're working on that...
@@ -221,13 +186,9 @@ public:
    * longitude (degrees), `Y` is latitude (degrees), and `Z` is height above the
    * ellipsoid (meters). Only valid if the placement type is Cartographic Origin
    * (i.e. Longitude / Latitude / Height).
-   *
-   * This converts the values to single-precision floating point values.
-   * The double-precision values can be accessed via the
-   * OriginLongitude, OriginLatitude and OriginHeight properties.
    */
   UFUNCTION(BlueprintCallable, Category = "Cesium")
-  FVector InaccurateGetGeoreferenceOriginLongitudeLatitudeHeight() const;
+  FVector GetGeoreferenceOriginLongitudeLatitudeHeight() const;
 
   /**
    * This aligns the specified longitude in degrees (x), latitude in
@@ -235,7 +196,7 @@ public:
    * origin. I.e. it moves the globe so that these coordinates exactly fall on
    * the origin.
    *
-   * When the WorldOriginCamera of this instance is currently contained
+   * When the SubLevelCamera of this instance is currently contained in
    * the bounds of a sublevel, then this call has no effect.
    */
   void SetGeoreferenceOriginLongitudeLatitudeHeight(
@@ -252,7 +213,7 @@ public:
    * Unreal's world origin. I.e. it moves the globe so that these coordinates
    * exactly fall on the origin.
    *
-   * When the WorldOriginCamera of this instance is currently contained
+   * When the SubLevelCamera of this instance is currently contained in
    * the bounds of a sublevel, then this call has no effect.
    */
   void SetGeoreferenceOriginEcef(const glm::dvec3& TargetEcef);
@@ -263,11 +224,11 @@ public:
    * origin. I.e. it moves the globe so that these coordinates exactly fall on
    * the origin.
    *
-   * When the WorldOriginCamera of this instance is currently contained
+   * When the SubLevelCamera of this instance is currently contained in
    * the bounds of a sublevel, then this call has no effect.
    */
   UFUNCTION(BlueprintCallable, Category = "Cesium")
-  void InaccurateSetGeoreferenceOriginLongitudeLatitudeHeight(
+  void SetGeoreferenceOriginLongitudeLatitudeHeight(
       const FVector& TargetLongitudeLatitudeHeight);
 
   /**
@@ -275,11 +236,11 @@ public:
    * Unreal's world origin. I.e. it moves the globe so that these coordinates
    * exactly fall on the origin.
    *
-   * When the WorldOriginCamera of this instance is currently contained
+   * When the SubLevelCamera of this instance is currently contained in
    * the bounds of a sublevel, then this call has no effect.
    */
   UFUNCTION(BlueprintCallable, Category = "Cesium")
-  void InaccurateSetGeoreferenceOriginEcef(const FVector& TargetEcef);
+  void SetGeoreferenceOriginEcef(const FVector& TargetEcef);
 
   /*
    * USEFUL CONVERSION FUNCTIONS
@@ -303,7 +264,7 @@ public:
    * TransformLongitudeLatitudeHeightToEcef can be used.
    */
   UFUNCTION(BlueprintCallable, Category = "Cesium")
-  FVector InaccurateTransformLongitudeLatitudeHeightToEcef(
+  FVector TransformLongitudeLatitudeHeightToEcef(
       const FVector& LongitudeLatitudeHeight) const;
 
   /**
@@ -318,14 +279,9 @@ public:
    * Transforms the given Earth-Centered, Earth-Fixed (ECEF) coordinates into
    * WGS84 longitude in degrees (x), latitude in degrees (y), and height above
    * the ellipsoid in meters (z).
-   *
-   * This function peforms the computation in single-precision. When using
-   * the C++ API, corresponding double-precision function
-   * TransformEcefToLongitudeLatitudeHeight can be used.
    */
   UFUNCTION(BlueprintCallable, Category = "Cesium")
-  FVector
-  InaccurateTransformEcefToLongitudeLatitudeHeight(const FVector& Ecef) const;
+  FVector TransformEcefToLongitudeLatitudeHeight(const FVector& Ecef) const;
 
   /**
    * Transforms the given longitude in degrees (x), latitude in
@@ -339,13 +295,9 @@ public:
    * Transforms the given longitude in degrees (x), latitude in
    * degrees (y), and height above the ellipsoid in meters (z) into Unreal world
    * coordinates (relative to the floating origin).
-   *
-   * This function peforms the computation in single-precision. When using
-   * the C++ API, corresponding double-precision function
-   * TransformLongitudeLatitudeHeightToUnreal can be used.
    */
   UFUNCTION(BlueprintCallable, Category = "Cesium")
-  FVector InaccurateTransformLongitudeLatitudeHeightToUnreal(
+  FVector TransformLongitudeLatitudeHeightToUnreal(
       const FVector& LongitudeLatitudeHeight) const;
 
   /**
@@ -360,14 +312,9 @@ public:
    * Transforms Unreal world coordinates (relative to the floating origin) into
    * longitude in degrees (x), latitude in degrees (y), and height above the
    * ellipsoid in meters (z).
-   *
-   * This function peforms the computation in single-precision. When using
-   * the C++ API, corresponding double-precision function
-   * TransformUnrealToLongitudeLatitudeHeight can be used.
    */
   UFUNCTION(BlueprintCallable, Category = "Cesium")
-  FVector InaccurateTransformUnrealToLongitudeLatitudeHeight(
-      const FVector& Unreal) const;
+  FVector TransformUnrealToLongitudeLatitudeHeight(const FVector& Unreal) const;
 
   /**
    * Transforms the given point from Earth-Centered, Earth-Fixed (ECEF) into
@@ -378,13 +325,9 @@ public:
   /**
    * Transforms the given point from Earth-Centered, Earth-Fixed (ECEF) into
    * Unreal relative world (relative to the floating origin).
-   *
-   * This function peforms the computation in single-precision. When using
-   * the C++ API, corresponding double-precision function
-   * TransformEcefToUnreal can be used.
    */
   UFUNCTION(BlueprintCallable, Category = "Cesium")
-  FVector InaccurateTransformEcefToUnreal(const FVector& Ecef) const;
+  FVector TransformEcefToUnreal(const FVector& Ecef) const;
 
   /**
    * Transforms the given point from Unreal relative world (relative to the
@@ -395,13 +338,9 @@ public:
   /**
    * Transforms the given point from Unreal relative world (relative to the
    * floating origin) to Earth-Centered, Earth-Fixed (ECEF).
-   *
-   * This function peforms the computation in single-precision. When using
-   * the C++ API, corresponding double-precision function
-   * TransformUnrealToEcef can be used.
    */
   UFUNCTION(BlueprintCallable, Category = "Cesium")
-  FVector InaccurateTransformUnrealToEcef(const FVector& Unreal) const;
+  FVector TransformUnrealToEcef(const FVector& Unreal) const;
 
   /**
    * Transforms a rotator from Unreal world to East-South-Up at the given
@@ -414,13 +353,9 @@ public:
   /**
    * Transforms a rotator from Unreal world to East-South-Up at the given
    * Unreal world location (relative to the floating origin).
-   *
-   * This function peforms the computation in single-precision. When using
-   * the C++ API, corresponding double-precision function
-   * TransformRotatorUnrealToEastSouthUp can be used.
    */
   UFUNCTION(BlueprintCallable, Category = "Cesium")
-  FRotator InaccurateTransformRotatorUnrealToEastSouthUp(
+  FRotator TransformRotatorUnrealToEastSouthUp(
       const FRotator& UnrealRotator,
       const FVector& UnrealLocation) const;
 
@@ -435,13 +370,9 @@ public:
   /**
    * Transforms a rotator from East-South-Up to Unreal world at the given
    * Unreal world location (relative to the floating origin).
-   *
-   * This function peforms the computation in single-precision. When using
-   * the C++ API, corresponding double-precision function
-   * TransformRotatorEastSouthUpToUnreal can be used.
    */
   UFUNCTION(BlueprintCallable, Category = "Cesium")
-  FRotator InaccurateTransformRotatorEastSouthUpToUnreal(
+  FRotator TransformRotatorEastSouthUpToUnreal(
       const FRotator& EsuRotator,
       const FVector& UnrealLocation) const;
 
@@ -458,13 +389,9 @@ public:
    * specified Unreal world location (relative to the floating
    * origin). The returned transformation works in Unreal's left-handed
    * coordinate system.
-   *
-   * This function peforms the computation in single-precision. When using
-   * the C++ API, corresponding double-precision function
-   * ComputeEastSouthUpToUnreal can be used.
    */
   UFUNCTION(BlueprintCallable, Category = "Cesium")
-  FMatrix InaccurateComputeEastSouthUpToUnreal(const FVector& Unreal) const;
+  FMatrix ComputeEastSouthUpToUnreal(const FVector& Unreal) const;
 
   /**
    * Computes the rotation matrix from the local East-North-Up to
@@ -475,13 +402,9 @@ public:
   /**
    * Computes the rotation matrix from the local East-North-Up to
    * Earth-Centered, Earth-Fixed (ECEF) at the specified ECEF location.
-   *
-   * This function peforms the computation in single-precision. When using
-   * the C++ API, corresponding double-precision function
-   * ComputeEastNorthUpToEcef can be used.
    */
   UFUNCTION(BlueprintCallable, Category = "Cesium")
-  FMatrix InaccurateComputeEastNorthUpToEcef(const FVector& Ecef) const;
+  FMatrix ComputeEastNorthUpToEcef(const FVector& Ecef) const;
 
   /**
    * @brief Computes the normal of the plane tangent to the surface of the
@@ -621,26 +544,12 @@ private:
    * @brief Updates the load state of sublevels.
    *
    * This checks all sublevels whether their load radius contains the
-   * `WorldOriginCamera`, in ECEF coordinates. The sublevels that
+   * `SubLevelCamera`, in ECEF coordinates. The sublevels that
    * contain the camera will be loaded. All others will be unloaded.
    *
    * @return Whether the camera is contained in *any* sublevel.
    */
   bool _updateSublevelState();
-
-  /**
-   * @brief Perform the origin-rebasing.
-   *
-   * If this actor is currently "in-game", and has an associated
-   * `WorldOriginCamera`, and the camera is further away from the origin than
-   * `MaximumWorldOriginDistanceFromCamera`, then this may set a new world
-   * origin by calling `GetWorld()->SetNewWorldOrigin` with a new position.
-   *
-   * This will only be done if origin rebasing is enabled via
-   * `KeepWorldOriginNearCamera`, and the actor is either *not* in a sublevel,
-   * or `OriginRebaseInsideSublevels` is enabled.
-   */
-  void _performOriginRebasing();
 
   /**
    * Updates _geoTransforms based on the current ellipsoid and center, and