+++ Source/CesiumRuntime/Public/CesiumMetadataValueType.h
@@ -16,6 +16,7 @@ enum class ECesiumMetadataBlueprintType : uint8 {
   Integer,
   Integer64,
   Float,
+  Float64,
   String,
   Array
 };