+++ Source/CesiumRuntime/Public/GlobeAwareDefaultPawn.h
@@ -142,8 +142,8 @@ public:
    */
   void FlyToLocationECEF(
       const glm::dvec3& ECEFDestination,
-      float YawAtDestination,
-      float PitchAtDestination,
+      double YawAtDestination,
+      double PitchAtDestination,
       bool CanInterruptByMoving);
 
   /**
@@ -155,10 +155,10 @@ public:
    * {@see FlyToGranularityDegrees}.
    */
   UFUNCTION(BlueprintCallable, Category = "Cesium")
-  void InaccurateFlyToLocationECEF(
+  void FlyToLocationECEF(
       const FVector& ECEFDestination,
-      float YawAtDestination,
-      float PitchAtDestination,
+      double YawAtDestination,
+      double PitchAtDestination,
       bool CanInterruptByMoving);
 
   /**
@@ -171,8 +171,8 @@ public:
    */
   void FlyToLocationLongitudeLatitudeHeight(
       const glm::dvec3& LongitudeLatitudeHeightDestination,
-      float YawAtDestination,
-      float PitchAtDestination,
+      double YawAtDestination,
+      double PitchAtDestination,
       bool CanInterruptByMoving);
 
   /**
@@ -184,10 +184,10 @@ public:
    * {@see FlyToDuration}, and {@see FlyToGranularityDegrees}.
    */
   UFUNCTION(BlueprintCallable, Category = "Cesium")
-  void InaccurateFlyToLocationLongitudeLatitudeHeight(
+  void FlyToLocationLongitudeLatitudeHeight(
       const FVector& LongitudeLatitudeHeightDestination,
-      float YawAtDestination,
-      float PitchAtDestination,
+      double YawAtDestination,
+      double PitchAtDestination,
       bool CanInterruptByMoving);
 
   virtual bool ShouldTickIfViewportsOnly() const override;
@@ -225,8 +225,8 @@ protected:
   UCesiumGlobeAnchorComponent* GlobeAnchor;
 
 private:
-  void _moveAlongViewAxis(EAxis::Type axis, float Val);
-  void _moveAlongVector(const FVector& axis, float Val);
+  void _moveAlongViewAxis(EAxis::Type axis, double Val);
+  void _moveAlongVector(const FVector& axis, double Val);
   void _interruptFlight();
 
   /**